Hoist up some of the fields from WorkflowMetaVellumDisplayOverrides to WorkflowMetaDisplay
